create an utils file for CursorTextField component
