+++ formulaire/frontend/src/components/CursorTextField/index.tsx
@@ -1,10 +1,11 @@
 import { TextField } from "@cgi-learning-hub/ui";
 import { FC } from "react";
-import { ICursorSlotProps, ICursorTextFieldProps } from "./types";
+import { ICursorTextFieldProps } from "./types";
 import { ComponentSize, ComponentVariant } from "~/core/style/themeProps";
 import { CursorTextFieldType } from "./enums";
 import { FORMULAIRE } from "~/core/constants";
 import { useTranslation } from "react-i18next";
+import { getInputSlotProps } from "./utils";
 
 export const CursorTextField: FC<ICursorTextFieldProps> = ({
   type,
@@ -15,46 +16,14 @@ export const CursorTextField: FC<ICursorTextFieldProps> = ({
 }) => {
   const { t } = useTranslation(FORMULAIRE);
   const isTypeNumber = type == CursorTextFieldType.NUMBER;
-
-  const getInputSlotProps = () => {
-    const slotProps: ICursorSlotProps = {
-      input: {
-        readOnly: !isCurrentEditingElement,
-        sx: {
-          pointerEvents: !isCurrentEditingElement ? "none" : "auto",
-          caretColor: !isCurrentEditingElement ? "transparent" : "auto",
-          width: isTypeNumber ? "100px" : "unset",
-        },
-      },
-    };
-
-    const commonInputProps = {
-      value: inputValue,
-    };
-
-    if (isTypeNumber) {
-      slotProps.htmlInput = {
-        ...commonInputProps,
-        step: stepValue,
-        inputMode: "numeric",
-        pattern: "[0-9]*",
-        sx: {
-          textAlign: "center",
-        },
-      };
-    } else {
-      slotProps.htmlInput = commonInputProps;
-    }
-
-    return slotProps;
-  };
+  const slotProps = getInputSlotProps(isCurrentEditingElement, isTypeNumber, inputValue, stepValue);
 
   return (
     <TextField
       type={type}
       variant={type == CursorTextFieldType.NUMBER ? ComponentVariant.OUTLINED : ComponentVariant.STANDARD}
       {...(isTypeNumber ? { size: ComponentSize.SMALL } : { placeholder: t("formulaire.question.label") })}
-      slotProps={getInputSlotProps()}
+      slotProps={slotProps}
       onChange={onChange}
     />
   );
